+++ src/main/java/bee/Bee.java
@@ -301,7 +301,7 @@ public class Bee {
      * @param tasks A list of task commands
      */
     public static void main(String... tasks) {
-        if (tasks.length == 0) tasks = new String[] {"native"};
+        if (tasks.length == 0) tasks = new String[] {"help"};
 
         System.exit(new Bee().execute(BeeOption.parse(tasks)));
     }